Move react app to ui directory
